refactoring: download buttons component
